Update README.md
Hi,

This is my first ever time creating a branch and making a commit using Git.  Please let me know if I did this correctly; I am excited to expand my knowledge in GitHub with all aspects of GIS.

I used TNTgis, more specifically, TNTMips as the main software for a geospatial image analysis course at the University of Denver a few years back.  This was part of a Master's Degree in GIS course curriculum.

Thank you for your time!
-Ethan